requirements.txt
absl-py==0.7.1
args==0.1.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
biwrap==0.1.6
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
click==6.7
clint==0.5.1
common==0.1.2
cycler==0.10.0
datasource==0.2.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.4
grpcio==1.19.0
h5py==2.9.0
hestia==0.3.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
marshmallow==3.0.0rc4
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.2
opencv-python==4.1.0.25
pandocfilters==1.4.2
parso==0.4.0
pathlib==1.0.1
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
Pillow-PIL==0.1.dev0
polyaxon-cli==0.4.4
polyaxon-client==0.4.4
polyaxon-deploy==0.4.4
polyaxon-helper==0.4.4
polyaxon-schemas==0.4.4
polystores==0.1.8
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.4.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
raven==6.7.0
requests==2.22.0
requests-toolbelt==0.8.0
rhea==0.4.5
scikit-image==0.15.0
scipy==1.1.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
tabulate==0.8.2
tb-nightly==1.14.0a20190301
tensorboard==1.12.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
